update retire, general lock update support

+++ contracts/fio.token/include/fio.token/fio.token.hpp
@@ -60,7 +60,7 @@ namespace eosio {
         void mintfio(const name &to, const uint64_t &amount);
 
         [[eosio::action]]
-        void retire(const int &quantity, const string &memo, const name &actor);
+        void retire(const int64_t &quantity, const string &memo, const name &actor);
 
         [[eosio::action]]
         void transfer(name from,
